news-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
decorator==5.2.1
defusedxml==0.7.1
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.3
django-redis==6.0.0
django-tinymce==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
executing==2.2.1
fastjsonschema==2.21.2
frozenlist==1.8.0
google-auth==2.49.0
gunicorn==21.2.0
http_ece==1.2.1
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
oauthlib==3.3.1
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.1
pipreqs==0.5.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-vapid==1.9.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pywebpush==2.3.0
pyzmq==27.1.0
redis==7.3.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.4
traitlets==5.14.3
tweepy==4.16.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
webencodings==0.5.1
yarg==0.1.9
yarl==1.23.0
boto3==1.34.0
django-storages==1.14.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
whitenoise==6.6.0
sentry-sdk[django]>=2,<3
